Fix data-layer audit findings: NaN serialization, overflow, save-race, zip-bomb OOM, dead code, docs
Addresses the core/data, core/domain, core/common batch of the 9-way adversarial
audit:

- CaptureEnvironment's NaN attitude/location sentinels now serialize: ProjectManager's
  Json sets allowSpecialFloatingPointValues = true, with a regression test that
  round-trips a project through the real save/load path with NaN fields (would have
  thrown before the fix).
- WallFeatureMap's points3d/descriptorsRows size check widened to Long so a hostile
  row count can't wrap the 32-bit multiply and slip past validation; matching native
  fix in GraffitiJNI.cpp's nativeRestoreWallFeatureMap (ptsLen != rows*3 was still
  32-bit three lines after the already-fixed rows*cols*elemSize check, feeding an
  unbounded reserve() outside any try/catch).
- Target capture in MainViewModel now routes through
  ProjectRepository.updateProject{} (the same atomic read-modify-write the editor's
  design-layer save and AR's wall-map save use) instead of a raw whole-object
  ProjectManager.saveProject, so it can no longer silently clobber (or be clobbered
  by) a concurrent writer.
- Zip import/spectator-load now stream each entry straight to a temp file instead of
  buffering it whole in a doubling ByteArrayOutputStream, so the size cap is actually
  protective against a zip bomb; duplicate entry names no longer leak a superseded
  temp file, and extraction always cleans up on the exception path too.
- Target images are now capped (30) and pruned (with file deletion) on capture,
  instead of growing a project's storage without bound.
- Removed ProjectManager.loadProject/LoadedProject: dead code with no caller
  anywhere; the "persist migration on full load" behavior two comments described
  never actually happened, since every real load went through the always-read-only
  loadProjectMetadata. Comments corrected instead of the function being wired up.
- Replaced ProjectRepositoryTest.kt's 13 tautological mock-returns-stub tests with
  real ProjectRepositoryImpl coverage of updateProject's atomic transform, including
  two genuine multi-thread concurrency tests.
- GraffitiProjectTest's "preserves all fields" now asserts full object equality
  instead of checking 4 of ~30 fields.
- docs/data_layer.md and docs/data_formats.md rewritten to match what
  ProjectManager actually reads/writes (plain-ZIP .gxr, no binary voxel-map format,
  correct core/common module paths) instead of a speculative format the two docs
  disagreed with each other about.

Skipped: the co-op collab module's own zip-extraction path (collab/, out of scope);
schema-version field (explicitly deprioritized in the task).

+++ core/common/src/main/java/com/hereliesaz/graffitixr/common/model/WallFeatureMap.kt
@@ -47,7 +47,11 @@ data class WallFeatureMap(
         // Defensive: the parallel arrays must agree so the native restore (Phase 2) can trust
         // them and never index out of bounds. Empty optional arrays are allowed.
         require(descriptorsRows >= 0 && descriptorsCols >= 0) { "descriptor dims must be non-negative" }
-        require(points3d.size == descriptorsRows * 3) {
+        // Widened to Long: a crafted/corrupt descriptorsRows (e.g. from an imported .gxr) can be
+        // large enough that `descriptorsRows * 3` wraps a 32-bit Int to a small value, which would
+        // make this guard satisfiable against the wrapped result instead of the real one — the same
+        // class of bug fixed on the native side (see GraffitiJNI.cpp's 64-bit row/col/elemSize checks).
+        require(points3d.size.toLong() == descriptorsRows.toLong() * 3L) {
             "points3d (${points3d.size}) must be 3 * descriptorsRows ($descriptorsRows)"
         }
         require(descriptorsRows == 0 || descriptorsData.size % descriptorsRows == 0) {
